Initial implementation of periodic jobs
The core idea is that we use a sorted set in redis to keep a PeriodicJob
json string that contains a cron string plus enough data to enqueue a
job with some args. The sorted set is sorted by the unix timestamp of the
next scheduled time for the periodic job.

When a sorted periodic jobs score passes `now()`, a process will pull the
periodic job json, find the next timestamp the job should run, and use ZADD
to update the periodic job in redis. Redis will tell us if we changed the
value. If so, the process will enqueue a job.