Update readme with details re: num_workers config and separate push/fetch pools
The `sidekiq::Processor` was recently updated to allow configuring the
number of worker tasks it spawns. Add a section for this to the readme.

Also, add a "Best practices" section, and add the recommendation to use
a separate connection pool for enqueuing vs fetching from Redis as
discussed in https://github.com/film42/sidekiq-rs/issues/43.