Attempt 1: Add a new field to track expected enqueue time
This adds an extra field for all schedule later jobs which is kind of
expensive considering the "score" is already in redis and easy to pull
out. I just don't love dangling a random metadata field on the Job
object. I'm also not sure if this warrants a new object in the
middleware call stack for each job.

So I'm looking for a way to minimize the cost for presenting this to the
end user without taxing everyone else who doesn't really care about it.
I do want it to be available though.