Refactor retry to be true/false or N; backward compatible
This changes JsonValue to RetryOpts which is an enum for Yes, Never, and
Max(n) to support the sidekiq bool|int typing. The public builder
interfaces for opts have now been moved to Into<RetryOpts> which will
continue to work for true/false like before but also a numeric value.

Also... I found that I hadn't impl'd the RetryMiddleware to respect a
retry: false option, so that's been fixed. I also added max_retries to
be plucked from the job if set and will be used in favor of whatever the
value was for the worker since they can be different.

TODO: Need to add a test for the new retry middleware logic. That
shouldn't have been a bug in the first place.