Only enqueue a sched job if the worker removed it
It was reported in https://github.com/film42/sidekiq-rs/pull/50 that we
have a race-condition with this scheduled enqueuer code. I think the
issue was that I was _assuming_ that redis would coerce a 0 to false and
a 1 to true, but  I don't think that's really the case. (Still need to
double check that).

Better yet, we should add a new test to prove this.

NOTE: I also updated the `n` reported by this method to only count the
`n` if it actually enqueued a job.

NOTE 2: I think it's best if we bring the enqueue batch size down from
100 to 10.