feat: Enable spawning workers dedicated to a specific queue
In a large application, it's possible for there to be a lot of queues,
each of which with many items. In addition, some queues may have a lot
of items, while others may only rarely have any items. These two factors
can lead to starvation of the rarely used queues if they get blocked
behind a large number of items in other more heavily used queues.

Add support for spawing workers that are dedicated to a specific queue.
This way, users can specify dedicated workers for their heavily used
queues, and the rarely used queues can continue using a shared set of
workers.